+++ test/jsonapi_client_test.exs
@@ -2,9 +2,11 @@ defmodule JsonApiClientTest do
   use ExUnit.Case
   doctest JsonApiClient, import: true
 
+  import Mock
   import JsonApiClient
   import JsonApiClient.Request
   alias JsonApiClient.{Request, Resource, Response, RequestError}
+  alias JsonApiClient.HTTPClient.HTTPoison
 
   setup do
     bypass = Bypass.open
@@ -12,150 +14,144 @@ defmodule JsonApiClientTest do
     {:ok, bypass: bypass, url: "http://localhost:#{bypass.port}"}
   end
 
-  test "includes status and headers from the HTTP response", context do
-    Bypass.expect context.bypass, "GET", "/articles/123", fn conn ->
-      conn
-      |> Plug.Conn.resp(200, "")
-      |> Plug.Conn.put_resp_header("X-Test-Header", "42")
+  test "return HTTP Backend response", context do
+    with_mock HTTPoison, [], [request: fn(_, _, _, _, _) -> {:ok, %Response{status: 201, doc: "foo"}} end] do
+      assert {:ok, %Response{status: 201, doc: "foo"}} = fetch Request.new(context.url <> "/articles/123")
     end
-
-    {:ok, response} = fetch Request.new(context.url <> "/articles/123")
-
-    assert response.status == 200
-    assert Enum.member?(response.headers, {"X-Test-Header", "42"})
   end
 
-  test "get a resource", context do
-    doc = single_resource_doc()
-    Bypass.expect context.bypass, "GET", "/articles/123", fn conn ->
-      assert_has_json_api_headers(conn)
-      Plug.Conn.resp(conn, 200, Poison.encode! doc)
+  test "add json api headers", context do
+    with_mock HTTPoison, [], [
+      request: fn(_, _, _, headers, _) ->
+          assert_has_json_api_headers(headers)
+          {:ok, %Response{status: 201}}
+        end
+      ] do
+      fetch Request.new(context.url <> "/articles/123")
     end
-
-    assert {:ok, %Response{status: 200, doc: ^doc}} = Request.new(context.url <> "/articles")
-    |> id("123")
-    |> method(:get)
-    |> execute
-
-    assert {:ok, %Response{status: 200, doc: ^doc}} = Request.new(context.url <> "/articles")
-    |> id("123")
-    |> fetch
   end
 
   test "set user agent with user suffix", context do
     Mix.Config.persist(json_api_client: [user_agent_suffix: "my_sufix"])
-    Bypass.expect context.bypass, "GET", "/articles/123", fn conn ->
-      assert Keyword.get(get_headers(conn), :"user-agent") == "json_api_client/" <> Mix.Project.config[:version] <> "/my_sufix"
-      Plug.Conn.resp(conn, 200, Poison.encode! single_resource_doc())
+    with_mock HTTPoison, [], [
+      request: fn(_, _, _, headers, _) ->
+          assert Keyword.get(to_atoms_keys(headers), :"User-Agent") == "json_api_client/" <> Mix.Project.config[:version] <> "/my_sufix"
+          {:ok, %Response{status: 201}}
+        end
+      ] do
+      fetch Request.new(context.url <> "/articles/123")
+      Mix.Config.persist(json_api_client: [user_agent_suffix: Mix.Project.config[:app]])
     end
-    Request.new(context.url <> "/articles") |> id("123") |> method(:get) |> execute
-    Mix.Config.persist(json_api_client: [user_agent_suffix: Mix.Project.config[:app]])
   end
 
   test "get a list of resources", context do
     doc = multiple_resource_doc()
-    Bypass.expect context.bypass, fn conn ->
-      conn = Plug.Conn.fetch_query_params(conn)
-      assert %{
-        "fields" => %{
-          "articles" => "title,topic",
-          "authors" => "first-name,last-name,twitter",
-        },
-        "include" => "author",
-        "sort" => "id",
-        "page" => %{"size" => "10", "number" => "1"},
-        "filter" => %{"published" => "true"},
-        "custom1" => "1",
-        "custom2" => "2",
-      } = conn.query_params
-      assert_has_json_api_headers(conn)
-      Plug.Conn.resp(conn, 200, Poison.encode! doc)
-    end
+    with_mock HTTPoison, [], [
+      request: fn(:get, url, _, headers, options) ->
+          assert url == context.url <> "/articles"
+          assert_has_json_api_headers(headers)
+          assert [
+            {"custom1", "1"},
+            {"custom2", "2"},
+            {"fields[articles]", "title,topic"},
+            {"fields[authors]", "first-name,last-name,twitter"},
+            {"filter[published]", "true"},
+            {"include", "author"},
+            {"page[size]", "10"},
+            {"page[number]", "1"},
+            {"sort", "id"}] == Keyword.get(options, :params)
 
-    assert {:ok, %Response{status: 200, doc: ^doc}} = Request.new(context.url <> "/articles")
-    |> fields(articles: "title,topic", authors: "first-name,last-name,twitter")
-    |> include(:author)
-    |> sort(:id)
-    |> page(size: 10, number: 1)
-    |> filter(published: true)
-    |> params(custom1: 1, custom2: 2)
-    |> fetch
+          {:ok, %Response{status: 200, doc: doc}}
+        end
+      ] do
+      assert {:ok, %Response{status: 200, doc: ^doc}} = Request.new(context.url <> "/articles")
+      |> fields(articles: "title,topic", authors: "first-name,last-name,twitter")
+      |> include(:author)
+      |> sort(:id)
+      |> page(size: 10, number: 1)
+      |> filter(published: true)
+      |> params(custom1: 1, custom2: 2)
+      |> fetch
+    end
   end
 
   test "delete a resource", context do
-    Bypass.expect context.bypass, "DELETE", "/articles/123", fn conn ->
-      assert_has_json_api_headers(conn)
-      Plug.Conn.resp(conn, 204, "")
+    with_mock HTTPoison, [], [
+      request: fn(:delete, url, _, headers, _) ->
+          assert url == context.url <> "/articles/123"
+          assert_has_json_api_headers(headers)
+          {:ok, %Response{status: 204}}
+        end
+      ] do
+
+      assert {:ok, %Response{status: 204, doc: nil}} = Request.new(context.url <> "/articles")
+      |> id("123")
+      |> delete
     end
-
-    assert {:ok, %Response{status: 204, doc: nil}} = Request.new(context.url)
-    |> resource(%Resource{type: "articles", id: "123"})
-    |> delete
-
-    assert {:ok, %Response{status: 204, doc: nil}} = Request.new(context.url <> "/articles")
-    |> id("123")
-    |> delete
   end
 
   test "create a resource", context do
     doc = single_resource_doc()
-    Bypass.expect context.bypass, "POST", "/articles", fn conn ->
-      assert_has_json_api_headers(conn)
+    with_mock HTTPoison, [], [
+      request: fn(:post, url, body, headers, _) ->
+          assert url == context.url <> "/articles"
+          assert_has_json_api_headers(headers)
+          assert %{
+            "data" => %{
+              "type" => "articles",
+              "attributes" => %{
+                "title" => "JSON API paints my bikeshed!",
+              },
+            }
+          } = Poison.decode! body
 
-      {:ok, body, conn} = Plug.Conn.read_body(conn)
-      assert %{
-        "data" => %{
-          "type" => "articles",
-          "attributes" => %{
-            "title" => "JSON API paints my bikeshed!",
-          },
+          {:ok, %Response{status: 201, doc: doc}}
+        end
+      ] do
+
+      new_article = %Resource{
+        type: "articles",
+        attributes: %{
+          title: "JSON API paints my bikeshed!",
         }
-      } = Poison.decode! body
-
-      Plug.Conn.resp(conn, 201, Poison.encode! doc)
-    end
-
-    new_article = %Resource{
-      type: "articles",
-      attributes: %{
-        title: "JSON API paints my bikeshed!",
       }
-    }
 
-    assert {:ok, %Response{status: 201, doc: ^doc}} = Request.new(context.url)
-    |> resource(new_article)
-    |> create
+      assert {:ok, %Response{status: 201, doc: ^doc}} = Request.new(context.url)
+      |> resource(new_article)
+      |> create
+    end
   end
 
   test "update a resource", context do
     doc = single_resource_doc()
-    Bypass.expect context.bypass, "PATCH", "/articles/123", fn conn ->
-      assert_has_json_api_headers(conn)
+    with_mock HTTPoison, [], [
+      request: fn(:patch, url, body, headers, _) ->
+          assert url == context.url <> "/articles"
+          assert_has_json_api_headers(headers)
+          assert %{
+            "data" => %{
+              "type" => "articles",
+              "attributes" => %{
+                "title" => "JSON API paints my bikeshed!",
+              },
+            }
+          } = Poison.decode! body
 
-      {:ok, body, conn} = Plug.Conn.read_body(conn)
-      assert %{
-        "data" => %{
-          "type" => "articles",
-          "attributes" => %{
-            "title" => "JSON API paints my bikeshed!",
-          },
+          {:ok, %Response{status: 201, doc: doc}}
+        end
+      ] do
+
+      new_article = %Resource{
+        type: "articles",
+        attributes: %{
+          title: "JSON API paints my bikeshed!",
         }
-      } = Poison.decode! body
-
-      Plug.Conn.resp(conn, 200, Poison.encode! doc)
-    end
-
-    new_article = %Resource{
-      type: "articles",
-      id: "123",
-      attributes: %{
-        title: "JSON API paints my bikeshed!",
       }
-    }
 
-    assert {:ok, %Response{status: 200, doc: ^doc}} = Request.new(context.url)
-    |> resource(new_article)
-    |> update
+      assert {:ok, %Response{status: 201, doc: ^doc}} = Request.new(context.url)
+      |> resource(new_article)
+      |> update
+    end
   end
 
   describe "Error Contidions" do
@@ -312,15 +308,18 @@ defmodule JsonApiClientTest do
     }
   end
 
-  def get_headers(conn) do
-    for {name, value} <- conn.req_headers, do: {String.to_atom(name), value}
+  def to_map(keyword) do
+    for [key, val] <- Enum.chunk(keyword, 2), do: {key, val}
   end
 
-  def assert_has_json_api_headers(conn) do
-    headers = get_headers(conn)
+  def to_atoms_keys(headers) do
+    for {name, value} <- headers, do: {String.to_atom(name), value}
+  end
 
-    assert Keyword.get(headers, :accept) == "application/vnd.api+json"
-    assert Keyword.get(headers, :"content-type") == "application/vnd.api+json"
-    assert Keyword.get(headers, :"user-agent") == "json_api_client/" <> Mix.Project.config[:version] <> "/#{Mix.Project.config[:app]}"
+  def assert_has_json_api_headers(headers) do
+    headers = to_atoms_keys(headers)
+    assert Keyword.get(headers, :"Accept") == "application/vnd.api+json"
+    assert Keyword.get(headers, :"Content-Type") == "application/vnd.api+json"
+    assert Keyword.get(headers, :"User-Agent") == "json_api_client/" <> Mix.Project.config[:version] <> "/#{Mix.Project.config[:app]}"
   end
 end