+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Renamed client_name to user_agent_suffix
 - Use package name as prefix instead of ExApiClient.
 - set Mix.Project.config[:app] as a default value for user_agent_suffix
+- added possibility to configure HTTP Client Backend 
 
 ## [1.0.0] - 2017-10-17
 ### Added